feat: log CONNECT error response body (#139)
Connected to https://github.com/apify/apify-proxy/issues/823.

On an internal proxy error (such as trying to access an invalid or
unresolvable URL - e.g. `https://example.comundefined`), the proxy
returns a response with the error message. `got-scraping` only logs the
length of the message, which is quite useless for most applications.